refactored API error handling

+++ apps/ds-api/src/routes/actions/custom/delete.ts
@@ -1,7 +1,7 @@
-import express, { Request, Response } from 'express';
+import express, { Request, Response, NextFunction } from 'express';
 import { db } from '../../../services/db';
 import { param } from 'express-validator';
-import { validateRequest } from '@devouringscripture/common';
+import { validateRequest, NotFoundError, DatabaseError, CustomError } from '@devouringscripture/common';
 
 const router = express.Router();
 
@@ -9,15 +9,22 @@ router.delete(
   '/:id',
   [param('id').isUUID().withMessage('Valid ID required')],
   validateRequest,
-  async (req: Request, res: Response) => {
+  async (req: Request, res: Response, next: NextFunction) => {
     console.log(`Delete custom action called for ${req.params.id}`);
 
     try {
       const indexOfItem = db.getIndex('/actions/custom', req.params.id);
+      if (indexOfItem < 0) {
+        throw new NotFoundError('custom action');
+      }
       db.delete(`/actions/custom[${indexOfItem}]`);
       res.status(200).send('Item deleted');
     } catch (err) {
-      res.status(500).send('Error deleting item');
+      if (err instanceof CustomError) {
+        return next(err);
+      }
+
+      return next(new DatabaseError('deleteCustomAction'));
     }
   }
 );
