+++ apps/ds-api/src/routes/read/passages/delete.ts
@@ -1,6 +1,6 @@
-import express, { Request, Response } from 'express';
+import express, { Request, Response, NextFunction } from 'express';
 import { param } from 'express-validator';
-import { validateRequest } from '@devouringscripture/common';
+import { validateRequest, NotFoundError, DatabaseError, CustomError } from '@devouringscripture/common';
 import { db } from '../../../services/db';
 
 const router = express.Router();
@@ -9,15 +9,18 @@ router.delete(
   '/:id',
   [param('id').isUUID().withMessage('ID required')],
   validateRequest,
-  async (req: Request, res: Response) => {
+  async (req: Request, res: Response, next: NextFunction) => {
     console.log(`Delete current read item with id ${req.params.id}`);
 
     try {
       const indexOfItem = db.getIndex('/passages', req.params.id);
+      if (indexOfItem < 0) {
+        throw new NotFoundError('Passage item');
+      }
       db.delete(`/passages[${indexOfItem}]`);
       res.send('Item removed');
     } catch (err) {
-      res.status(500).send('Error removing item');
+      return next(err instanceof CustomError ? err : new DatabaseError('deletePassage'));
     }
   }
 );
